% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdatasaver-package.R
\docType{package}
\name{ggdatasaver-package}
\alias{ggdatasaver}
\alias{ggdatasaver-package}
\title{ggdatasaver: Automatically save data associated with a 'ggplot2' plot}
\description{
Automatically create a file with the data used to create a 'ggplot2' plot. Each layer's data is saved in its own csv file inside a compressed file. These files can then be published as supplemental material so that other people can re-do your plots without the need of digitising them. Because the only data that is saved is the one that is shown in the plot, there are fewer privacy concerns.
}
\author{
\strong{Maintainer}: Elio Campitelli \email{elio.campitelli@cima.fcen.uba.ar} (\href{https://orcid.org/0000-0002-7742-9230}{ORCID})

}
\keyword{internal}
