% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data-dir-set.R
\name{plot_data_dir_set}
\alias{plot_data_dir_set}
\title{Set the data directory}
\usage{
plot_data_dir_set(dir = NULL)
}
\arguments{
\item{dir}{A directory where to save plot data. If NULL,
plot data will not be saved.

This functions sets the \code{plot_data_dir} chunk option in knitr.
You can get the same behaviour with \code{knitr::opts_chunk$set(plot_data_dir = dir)}.
However, if you use that code, you'll need to call \code{library(ggdatasaver)}
explicitly for the package to work.}
}
\description{
Set the directory where data assocaited with plots
will be saved.
}
