% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-plot-data.R
\name{save_plot_data}
\alias{save_plot_data}
\title{Saves data associated with a ggplot plot}
\usage{
save_plot_data(plot, name = "plot", dir = ".")
}
\arguments{
\item{plot}{A ggplot2 plot.}

\item{name}{The name used as file name.}

\item{dir}{Directory where to put the file. If NUL

In general, you don't need to call this function
explicitly, since it will be used automatically by knitr
if the plot data directory is set (see \link{plot_data_dir_set}).

Paraphrasing Voltaire, if \code{dir} doesn't exist,
it will be created. If the file already exists, it will
be overwritten without a warning. This is because this
function is supposed to run with knitr every time a document
is knit and will update the data to the latest version.}
}
\description{
Takes a ggplot and saves a zip file with a single csv
file for each layer of the plot with the data used by
said layer
}
